requirements.txt
pytest
pytest-cov
pandas
numpy
plotly
biopython